pyproject.toml
[tool.poetry]
name = "libclang-example"
version = "0.1.0"
description = ""
authors = ["Hiroki Ishioka <ishioka0222@gmail.com>"]
readme = "README.md"
packages = [{ include = "libclang_example", from = "src" }]


[tool.poetry.scripts]
main = "libclang_example.main:main"


[tool.poetry.dependencies]
python = "^3.11"
libclang = "^16.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
